onnx_model_serving/requirements.txt
matplotlib==3.3.3
numpy==1.19.5
requests==2.25.1
onnxruntime==1.6.0
onnx==1.8.1
torch==1.7.1+cu110
Pillow==8.1.0
scikit_learn==0.24.1
